requirements.txt
requests
simplejson
Django==1.6
djangorestframework==2.3.8
South==0.8.2
django-simple-captcha==0.4.0
gunicorn==18.0
hbase-thrift==0.20.4
thrift==0.9.1
xlrd==0.9.2
lxml==3.3.3
docx2html==0.2.3
